tests/wpt/css-tests/selectors-3_dev/xhtml1print/css3-selectors-lang-022.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html lang="en" xmlns="http://www.w3.org/1999/xhtml"><head>
<meta charset="utf-8" />
<title>[lang|="es"], lang="es" on parent</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida" />
<link href="http://www.w3.org/TR/css3-selectors/#attribute-selectors" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type="text/css">
.test div { width: 50px; }
#box[lang|='es'] { width: 100px; }
</style>
</head>
<body>



<div lang="es" class="test"><div id="box">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, "A lang|= value that matches an identical lang attribute value on the parent element will NOT produce styling.");
</script>

<div id="log"></div>



</body></html>
